Remove loader and util from respondentTrack model and move transformer
